Separate out in-memory content manager impl, and bundle the file impl with the REST server
