+++ backend/scribble-web-rest/pom.xml
@@ -58,6 +58,10 @@
       <groupId>org.scribble.tools.web</groupId>
       <artifactId>scribble-web-impl</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.scribble.tools.web</groupId>
+      <artifactId>scribble-cm-filesystem</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>junit</groupId>
